+++ server.h
@@ -1,6 +1,8 @@
 #ifndef JOURNALD__SERVER__H__
 #define JOURNALD__SERVER__H__
 
+#include <uint32.h>
+
 #define IDENTSIZE 1024
 #define CBUFSIZE 8192
 
@@ -10,14 +12,14 @@ struct connection
   int mode;
   int state;
   int count;
-  unsigned long length;
-  unsigned long ident_len;
-  unsigned long buf_length;
+  uint32 length;
+  uint32 ident_len;
+  uint32 buf_length;
   int wrote_ident;
   int ok;
-  unsigned long total;
-  unsigned long records;
-  unsigned long number;
+  uint32 total;
+  uint32 records;
+  uint32 number;
   
   char ident[IDENTSIZE];
   char buf[CBUFSIZE];
@@ -28,7 +30,7 @@ extern connection* connections;
 extern unsigned opt_connections;
 
 extern void die(const char* msg);
-extern void handle_data(connection* con, char* data, unsigned long size);
+extern void handle_data(connection* con, char* data, uint32 size);
 extern int open_journal(const char* filename);
 extern int write_record(connection* con, int final, int abort);
 extern int sync_records(void);
